+++ Automate/Framework/Machines/Buildings/ShippingBinMachine.cs
@@ -1,16 +1,18 @@
 using System.Linq;
+using Microsoft.Xna.Framework;
 using StardewValley;
+using StardewValley.Buildings;
 using SObject = StardewValley.Object;
 
 namespace Pathoschild.Stardew.Automate.Framework.Machines.Buildings
 {
     /// <summary>A shipping bin that accepts input and provides output.</summary>
-    internal class ShippingBinMachine : IMachine
+    internal class ShippingBinMachine : BaseMachine
     {
         /*********
         ** Properties
         *********/
-        /// <summary>The farm containing the shipping bin.</summary>
+        /// <summary>The farm to automate.</summary>
         private readonly Farm Farm;
 
 
@@ -19,19 +21,31 @@ namespace Pathoschild.Stardew.Automate.Framework.Machines.Buildings
         *********/
         /// <summary>Construct an instance.</summary>
         /// <param name="farm">The farm containing the shipping bin.</param>
-        public ShippingBinMachine(Farm farm)
+        /// <param name="tileArea">The tile area covered by the machine.</param>
+        public ShippingBinMachine(Farm farm, Rectangle tileArea)
+            : base(farm, tileArea)
+        {
+            this.Farm = farm;
+        }
+
+        /// <summary>Construct an instance.</summary>
+        /// <param name="bin">The constructed shipping bin.</param>
+        /// <param name="location">The location which contains the machine.</param>
+        /// <param name="farm">The farm which has the shipping bin data.</param>
+        public ShippingBinMachine(ShippingBin bin, GameLocation location, Farm farm)
+            : base(location, BaseMachine.GetTileAreaFor(bin))
         {
             this.Farm = farm;
         }
 
         /// <summary>Get the machine's processing state.</summary>
-        public MachineState GetState()
+        public override MachineState GetState()
         {
             return MachineState.Empty; // always accepts items
         }
 
         /// <summary>Get the output item.</summary>
-        public ITrackedStack GetOutput()
+        public override ITrackedStack GetOutput()
         {
             return null; // no output
         }
@@ -39,7 +53,7 @@ namespace Pathoschild.Stardew.Automate.Framework.Machines.Buildings
         /// <summary>Provide input to the machine.</summary>
         /// <param name="input">The available items.</param>
         /// <returns>Returns whether the machine started processing an item.</returns>
-        public bool SetInput(IStorage input)
+        public override bool SetInput(IStorage input)
         {
             ITrackedStack tracker = input.GetItems().Where(p => p.Sample is SObject obj && obj.canBeShipped()).Take(1).FirstOrDefault();
             if (tracker != null)