+++ Automate/Framework/Machines/Objects/OilMakerMachine.cs
@@ -1,10 +1,11 @@
 using Microsoft.Xna.Framework;
+using StardewValley;
 using SObject = StardewValley.Object;
 
 namespace Pathoschild.Stardew.Automate.Framework.Machines.Objects
 {
     /// <summary>An oil maker that accepts input and provides output.</summary>
-    internal class OilMakerMachine : GenericMachine
+    internal class OilMakerMachine : GenericObjectMachine<SObject>
     {
         /*********
         ** Properties
@@ -51,8 +52,9 @@ namespace Pathoschild.Stardew.Automate.Framework.Machines.Objects
         *********/
         /// <summary>Construct an instance.</summary>
         /// <param name="machine">The underlying machine.</param>
-        public OilMakerMachine(SObject machine)
-            : base(machine) { }
+        /// <param name="location">The location containing the machine.</param>
+        public OilMakerMachine(SObject machine, GameLocation location)
+            : base(machine, location) { }
 
         /// <summary>Provide input to the machine.</summary>
         /// <param name="input">The available items.</param>