+++ Automate/Framework/Machines/TerrainFeatures/FruitTreeMachine.cs
@@ -1,3 +1,4 @@
+using Microsoft.Xna.Framework;
 using StardewValley;
 using StardewValley.TerrainFeatures;
 using SObject = StardewValley.Object;
@@ -6,40 +7,33 @@ namespace Pathoschild.Stardew.Automate.Framework.Machines.TerrainFeatures
 {
     /// <summary>A fruit tree machine that accepts input and provides output.</summary>
     /// <remarks>Derived from <see cref="FruitTree.shake"/>.</remarks>
-    internal class FruitTreeMachine : IMachine
+    internal class FruitTreeMachine : BaseMachine<FruitTree>
     {
-        /*********
-        ** Properties
-        *********/
-        /// <summary>The underlying fruit tree.</summary>
-        private readonly FruitTree Tree;
-
-
         /*********
         ** Public methods
         *********/
         /// <summary>Construct an instance.</summary>
         /// <param name="tree">The underlying fruit tree.</param>
-        public FruitTreeMachine(FruitTree tree)
-        {
-            this.Tree = tree;
-        }
+        /// <param name="location">The machine's in-game location.</param>
+        /// <param name="tile">The tree's tile position.</param>
+        public FruitTreeMachine(FruitTree tree, GameLocation location, Vector2 tile)
+            : base(tree, location, BaseMachine.GetTileAreaFor(tile)) { }
 
         /// <summary>Get the machine's processing state.</summary>
-        public MachineState GetState()
+        public override MachineState GetState()
         {
-            if (this.Tree.growthStage.Value < FruitTree.treeStage)
+            if (this.Machine.growthStage.Value < FruitTree.treeStage)
                 return MachineState.Disabled;
 
-            return this.Tree.fruitsOnTree.Value > 0
+            return this.Machine.fruitsOnTree.Value > 0
                 ? MachineState.Done
                 : MachineState.Processing;
         }
 
         /// <summary>Get the output item.</summary>
-        public ITrackedStack GetOutput()
+        public override ITrackedStack GetOutput()
         {
-            FruitTree tree = this.Tree;
+            FruitTree tree = this.Machine;
 
             // if struck by lightning => coal
             if (tree.struckByLightningCountdown.Value > 0)
@@ -59,7 +53,7 @@ namespace Pathoschild.Stardew.Automate.Framework.Machines.TerrainFeatures
         /// <summary>Provide input to the machine.</summary>
         /// <param name="input">The available items.</param>
         /// <returns>Returns whether the machine started processing an item.</returns>
-        public bool SetInput(IStorage input)
+        public override bool SetInput(IStorage input)
         {
             return false; // no input
         }
@@ -72,7 +66,7 @@ namespace Pathoschild.Stardew.Automate.Framework.Machines.TerrainFeatures
         /// <param name="item">The output item that was taken.</param>
         private void OnOutputReduced(Item item)
         {
-            this.Tree.fruitsOnTree.Value = item.Stack;
+            this.Machine.fruitsOnTree.Value = item.Stack;
         }
     }
 }