+++ Automate/IContainer.cs
@@ -1,5 +1,6 @@
 using System;
 using System.Collections.Generic;
+using Microsoft.Xna.Framework;
 using StardewValley;
 
 namespace Pathoschild.Stardew.Automate
@@ -12,6 +13,12 @@ namespace Pathoschild.Stardew.Automate
         *********/
         /// <summary>The container name (if any).</summary>
         string Name { get; }
+
+        /// <summary>The location which contains the container.</summary>
+        GameLocation Location { get; }
+
+        /// <summary>The tile area covered by the container.</summary>
+        Rectangle TileArea { get; }
 
 
         /*********