+++ Automate/release-notes.md
@@ -3,6 +3,7 @@
 # Release notes
 ## Upcoming release
 * Fixed fences not working as connectors.
+* Internal changes to support custom machines in an upcoming release.
 
 ## 1.10.6
 * Migrated verbose logs to SMAPI's verbose logging feature.